incoming Create flow with tests
